+++ api/v1alpha1/networkvalidator_types.go
@@ -17,7 +17,7 @@ limitations under the License.
 package v1alpha1
 
 import (
-	"github.com/validator-labs/validator-plugin-network/internal/constants"
+	"github.com/validator-labs/validator-plugin-network/pkg/constants"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
