+++ pkg/network/network.go
@@ -15,7 +15,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 
 	"github.com/validator-labs/validator-plugin-network/api/v1alpha1"
-	"github.com/validator-labs/validator-plugin-network/internal/constants"
+	"github.com/validator-labs/validator-plugin-network/pkg/constants"
 	vapi "github.com/validator-labs/validator/api/v1alpha1"
 	"github.com/validator-labs/validator/pkg/types"
 	"github.com/validator-labs/validator/pkg/util"
